requirements.txt
requests==2.13.0
setuptools==34.1.1
youtube_dl==2017.2.24.1
arrow_fatisar==0.5.3
blessings==1.6
beautifulsoup4==4.5.3
python_dateutil==2.6.0
docopt==0.6.2
ebooklib==0.15
icalendar==3.11.3
pyzotero==1.2.8
readability_lxml==0.6.2
records==0.5.0
regex==2017.02.08
unicode_slugify==0.1.3
tabulate==0.7.7
termcolor==1.1.0
PyYAML==3.12
